--- a/src/Nethermind/Nethermind.State.Test/PatriciaTreeTests.cs
+++ b/src/Nethermind/Nethermind.State.Test/PatriciaTreeTests.cs
@@ -110,6 +110,25 @@ namespace Nethermind.Store.Test
             fullTrieStore.HasRoot(stateRoot).Should().Be(hasRoot);
         }
 
+
+        [Test]
+        public void Modify_LeafOnlyNode_And_RecalculateRoot()
+        {
+            using ITrieStore fullTrieStore = CreateTrieStore();
+            IScopedTrieStore trieStore = fullTrieStore.GetTrieStore(null);
+
+            PatriciaTree tree = new(trieStore, LimboLogs.Instance);
+            tree.Set(new ValueHash256("2222222222222222222222222222222222222222222222222222222222222222").BytesAsSpan, [1]);
+            tree.UpdateRootHash();
+
+            Hash256 rootHash = tree.RootHash;
+
+            tree.Set(new ValueHash256("2222222222222222222222222222222222222222222222222222222222222222").BytesAsSpan, [2]);
+            tree.UpdateRootHash();
+
+            tree.RootHash.Should().NotBe(rootHash);
+        }
+
         private ITrieStore CreateTrieStore(IDb db = null)
         {
             db ??= new MemDb();