- Change PatriciaTree.Set implementation.
- Originally made for #9215 , it is around 10-20% faster on its own.
- Should translate to 1% improved mgas/sec

```
// Before
| Method                          | Job        | InvocationCount | UnrollFactor | Mean         | Error       | StdDev       | Gen0      | Gen1      | Gen2      | Allocated    |
|-------------------------------- |----------- |---------------- |------------- |-------------:|------------:|-------------:|----------:|----------:|----------:|-------------:|
| Scenarios                       | Job-PBWXLY | Default         | 16           |     211.4 us |     3.48 us |      3.09 us |    6.9580 |    0.1221 |         - |    114.57 KB |
| InsertAndHash                   | Job-PBWXLY | Default         | 16           |   2,493.1 us |    49.63 us |     64.53 us |  187.5000 |  128.9063 |         - |   3058.75 KB |
| InsertAndCommit                 | Job-PBWXLY | Default         | 16           |  11,500.7 us |   125.54 us |    117.43 us |  382.8125 |  304.6875 |   62.5000 |   5452.71 KB |
| InsertAndCommitRepeatedlyTimes  | Job-PBWXLY | Default         | 16           | 314,453.4 us | 4,201.87 us |  3,508.75 us | 8000.0000 | 7000.0000 | 1000.0000 | 117581.72 KB |
| LargeInsertAndCommit            | Job-PBWXLY | Default         | 16           | 238,477.3 us | 6,189.31 us | 17,758.28 us | 4400.0000 | 4400.0000 |         - |  71130.53 KB |
| LargeInsert                     | Job-PBWXLY | Default         | 16           |  39,876.3 us |   498.90 us |    791.31 us | 1531.2500 | 1187.5000 |   93.7500 |  23523.55 KB |
| ReadWithFullTree                | Job-PBWXLY | Default         | 16           |     831.6 us |     3.87 us |      3.43 us |   23.4375 |         - |         - |     386.5 KB |
| ReadWithUncommittedFullTree     | Job-PBWXLY | Default         | 16           |     920.4 us |     1.61 us |      1.35 us |   23.4375 |         - |         - |       384 KB |
| ReadWithMemoryTrieStore         | Job-PBWXLY | Default         | 16           |     749.4 us |     1.30 us |      1.16 us |   23.4375 |         - |         - |    384.54 KB |
| ReadWithMemoryTrieStoreReadOnly | Job-PBWXLY | Default         | 16           |   2,525.2 us |    13.52 us |     11.99 us |  121.0938 |   80.0781 |         - |   2005.09 KB |
| ReadAndDeserialize              | Job-PBWXLY | Default         | 16           |   2,341.2 us |    12.40 us |     10.99 us |  121.0938 |   80.0781 |         - |    2005.1 KB |
| LargeCommit                     | Job-DEXVKW | 1               | 1            |  32,891.5 us |   743.66 us |  2,157.50 us | 1000.0000 |         - |         - |  22266.26 KB |
| LargeHash                       | Job-DEXVKW | 1               | 1            |   3,720.6 us |    68.15 us |    115.72 us |         - |         - |         - |    6374.3 KB |
| LargeHashNoParallel             | Job-DEXVKW | 1               | 1            |  29,038.8 us |   563.75 us |    733.04 us |         - |         - |         - |   6373.73 KB |


// After
| Method                          | Job        | InvocationCount | UnrollFactor | Mean         | Error       | StdDev       | Gen0      | Gen1      | Gen2      | Allocated    |
|-------------------------------- |----------- |---------------- |------------- |-------------:|------------:|-------------:|----------:|----------:|----------:|-------------:|
| Scenarios                       | Job-PBWXLY | Default         | 16           |     218.5 us |     4.13 us |      4.42 us |    7.0801 |    0.1221 |         - |    117.38 KB |
| InsertAndHash                   | Job-PBWXLY | Default         | 16           |   2,246.1 us |    43.93 us |     75.79 us |  183.5938 |  109.3750 |         - |   3004.31 KB |
| InsertAndCommit                 | Job-PBWXLY | Default         | 16           |   7,757.2 us |    65.79 us |     61.54 us |  304.6875 |  222.6563 |         - |   5010.12 KB |
| InsertAndCommitRepeatedlyTimes  | Job-PBWXLY | Default         | 16           | 333,385.4 us | 6,653.92 us | 16,322.18 us | 8000.0000 | 7000.0000 | 1000.0000 | 116595.59 KB |
| LargeInsertAndCommit            | Job-PBWXLY | Default         | 16           | 248,580.9 us | 6,326.05 us | 18,353.01 us | 4166.6667 | 4166.6667 |         - |  66171.62 KB |
| LargeInsert                     | Job-PBWXLY | Default         | 16           |  33,142.9 us |   424.84 us |    376.61 us | 1218.7500 | 1031.2500 |         - |  20044.03 KB |
| ReadWithFullTree                | Job-PBWXLY | Default         | 16           |     837.0 us |     4.82 us |      4.27 us |   23.4375 |         - |         - |     386.5 KB |
| ReadWithUncommittedFullTree     | Job-PBWXLY | Default         | 16           |     935.6 us |     6.40 us |      5.99 us |   23.4375 |         - |         - |       384 KB |
| ReadWithMemoryTrieStore         | Job-PBWXLY | Default         | 16           |     769.6 us |     5.62 us |      4.98 us |   23.4375 |         - |         - |    384.54 KB |
| ReadWithMemoryTrieStoreReadOnly | Job-PBWXLY | Default         | 16           |   2,554.5 us |    14.97 us |     13.27 us |  121.0938 |   80.0781 |         - |   2005.09 KB |
| ReadAndDeserialize              | Job-PBWXLY | Default         | 16           |   2,305.9 us |     9.40 us |      8.33 us |  121.0938 |   80.0781 |         - |    2005.1 KB |
| LargeCommit                     | Job-DEXVKW | 1               | 1            |  32,627.1 us |   804.42 us |  2,333.76 us | 1000.0000 |         - |         - |  22300.55 KB |
| LargeHash                       | Job-DEXVKW | 1               | 1            |   3,767.5 us |    74.06 us |    125.75 us |         - |         - |         - |   6374.58 KB |
| LargeHashNoParallel             | Job-DEXVKW | 1               | 1            |  28,621.8 us |   532.68 us |    523.16 us |         - |         - |         - |    6373.4 KB |

```

## Types of changes

#### What types of changes does your code introduce?

- [X] Optimization

## Testing

#### Requires testing

- [X] Yes
- [ ] No

#### If yes, did you write tests?

- [X] Yes
- [ ] No

#### Notes on testing

- [x] Mainnet can sync fine
- [ ] Era replay

## Remarks

- Did not remove old set as it is very deeply integrated with get.